+++ sdk/resourcemanager/quota/armquota/fake/groupquotalimitsrequest_server.go
@@ -26,8 +26,8 @@ import (
 // GroupQuotaLimitsRequestServer is a fake server for instances of the armquota.GroupQuotaLimitsRequestClient type.
 type GroupQuotaLimitsRequestServer struct {
 	// BeginCreateOrUpdate is the fake for method GroupQuotaLimitsRequestClient.BeginCreateOrUpdate
-	// HTTP status codes to indicate success: http.StatusOK, http.StatusCreated
-	BeginCreateOrUpdate func(ctx context.Context, managementGroupID string, groupQuotaName string, resourceProviderName string, resourceName string, options *armquota.GroupQuotaLimitsRequestClientBeginCreateOrUpdateOptions) (resp azfake.PollerResponder[armquota.GroupQuotaLimitsRequestClientCreateOrUpdateResponse], errResp azfake.ErrorResponder)
+	// HTTP status codes to indicate success: http.StatusOK, http.StatusAccepted
+	BeginCreateOrUpdate func(ctx context.Context, managementGroupID string, groupQuotaName string, resourceProviderName string, options *armquota.GroupQuotaLimitsRequestClientBeginCreateOrUpdateOptions) (resp azfake.PollerResponder[armquota.GroupQuotaLimitsRequestClientCreateOrUpdateResponse], errResp azfake.ErrorResponder)
 
 	// Get is the fake for method GroupQuotaLimitsRequestClient.Get
 	// HTTP status codes to indicate success: http.StatusOK
@@ -39,7 +39,7 @@ type GroupQuotaLimitsRequestServer struct {
 
 	// BeginUpdate is the fake for method GroupQuotaLimitsRequestClient.BeginUpdate
 	// HTTP status codes to indicate success: http.StatusOK, http.StatusAccepted
-	BeginUpdate func(ctx context.Context, managementGroupID string, groupQuotaName string, resourceProviderName string, resourceName string, options *armquota.GroupQuotaLimitsRequestClientBeginUpdateOptions) (resp azfake.PollerResponder[armquota.GroupQuotaLimitsRequestClientUpdateResponse], errResp azfake.ErrorResponder)
+	BeginUpdate func(ctx context.Context, managementGroupID string, groupQuotaName string, resourceProviderName string, options *armquota.GroupQuotaLimitsRequestClientBeginUpdateOptions) (resp azfake.PollerResponder[armquota.GroupQuotaLimitsRequestClientUpdateResponse], errResp azfake.ErrorResponder)
 }
 
 // NewGroupQuotaLimitsRequestServerTransport creates a new instance of GroupQuotaLimitsRequestServerTransport with the provided implementation.
@@ -100,10 +100,10 @@ func (g *GroupQuotaLimitsRequestServerTransport) dispatchBeginCreateOrUpdate(req
 	}
 	beginCreateOrUpdate := g.beginCreateOrUpdate.get(req)
 	if beginCreateOrUpdate == nil {
-		const regexStr = `/providers/Microsoft\.Management/managementGroups/(?P<managementGroupId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.Quota/groupQuotas/(?P<groupQuotaName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceProviders/(?P<resourceProviderName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/groupQuotaRequests/(?P<resourceName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)`
+		const regexStr = `/providers/Microsoft\.Management/managementGroups/(?P<managementGroupId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.Quota/groupQuotas/(?P<groupQuotaName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceProviders/(?P<resourceProviderName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/createLimitRequest`
 		regex := regexp.MustCompile(regexStr)
 		matches := regex.FindStringSubmatch(req.URL.EscapedPath())
-		if matches == nil || len(matches) < 4 {
+		if matches == nil || len(matches) < 3 {
 			return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
 		}
 		body, err := server.UnmarshalRequestAsJSON[armquota.SubmittedResourceRequestStatus](req)
@@ -122,17 +122,13 @@ func (g *GroupQuotaLimitsRequestServerTransport) dispatchBeginCreateOrUpdate(req
 		if err != nil {
 			return nil, err
 		}
-		resourceNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("resourceName")])
-		if err != nil {
-			return nil, err
-		}
 		var options *armquota.GroupQuotaLimitsRequestClientBeginCreateOrUpdateOptions
 		if !reflect.ValueOf(body).IsZero() {
 			options = &armquota.GroupQuotaLimitsRequestClientBeginCreateOrUpdateOptions{
 				GroupQuotaRequest: &body,
 			}
 		}
-		respr, errRespr := g.srv.BeginCreateOrUpdate(req.Context(), managementGroupIDParam, groupQuotaNameParam, resourceProviderNameParam, resourceNameParam, options)
+		respr, errRespr := g.srv.BeginCreateOrUpdate(req.Context(), managementGroupIDParam, groupQuotaNameParam, resourceProviderNameParam, options)
 		if respErr := server.GetError(errRespr, req); respErr != nil {
 			return nil, respErr
 		}
@@ -145,9 +141,9 @@ func (g *GroupQuotaLimitsRequestServerTransport) dispatchBeginCreateOrUpdate(req
 		return nil, err
 	}
 
-	if !contains([]int{http.StatusOK, http.StatusCreated}, resp.StatusCode) {
+	if !contains([]int{http.StatusOK, http.StatusAccepted}, resp.StatusCode) {
 		g.beginCreateOrUpdate.remove(req)
-		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusCreated", resp.StatusCode)}
+		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK, http.StatusAccepted", resp.StatusCode)}
 	}
 	if !server.PollerResponderMore(beginCreateOrUpdate) {
 		g.beginCreateOrUpdate.remove(req)
@@ -249,10 +245,10 @@ func (g *GroupQuotaLimitsRequestServerTransport) dispatchBeginUpdate(req *http.R
 	}
 	beginUpdate := g.beginUpdate.get(req)
 	if beginUpdate == nil {
-		const regexStr = `/providers/Microsoft\.Management/managementGroups/(?P<managementGroupId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.Quota/groupQuotas/(?P<groupQuotaName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceProviders/(?P<resourceProviderName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/groupQuotaRequests/(?P<resourceName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)`
+		const regexStr = `/providers/Microsoft\.Management/managementGroups/(?P<managementGroupId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.Quota/groupQuotas/(?P<groupQuotaName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/resourceProviders/(?P<resourceProviderName>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/createLimitRequest`
 		regex := regexp.MustCompile(regexStr)
 		matches := regex.FindStringSubmatch(req.URL.EscapedPath())
-		if matches == nil || len(matches) < 4 {
+		if matches == nil || len(matches) < 3 {
 			return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
 		}
 		body, err := server.UnmarshalRequestAsJSON[armquota.SubmittedResourceRequestStatus](req)
@@ -271,17 +267,13 @@ func (g *GroupQuotaLimitsRequestServerTransport) dispatchBeginUpdate(req *http.R
 		if err != nil {
 			return nil, err
 		}
-		resourceNameParam, err := url.PathUnescape(matches[regex.SubexpIndex("resourceName")])
-		if err != nil {
-			return nil, err
-		}
 		var options *armquota.GroupQuotaLimitsRequestClientBeginUpdateOptions
 		if !reflect.ValueOf(body).IsZero() {
 			options = &armquota.GroupQuotaLimitsRequestClientBeginUpdateOptions{
 				GroupQuotaRequest: &body,
 			}
 		}
-		respr, errRespr := g.srv.BeginUpdate(req.Context(), managementGroupIDParam, groupQuotaNameParam, resourceProviderNameParam, resourceNameParam, options)
+		respr, errRespr := g.srv.BeginUpdate(req.Context(), managementGroupIDParam, groupQuotaNameParam, resourceProviderNameParam, options)
 		if respErr := server.GetError(errRespr, req); respErr != nil {
 			return nil, respErr
 		}